+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Internal "MemoryMapManager" will make datatable more robust when opening a
   frame with many columns on Linux systems. In particular, error 12 "not enough
   memory" should become much more rare now.
+- Number of threads used by fread can now be controlled via parameter `nthreads`.
 
 #### Changed
 - When writing "round" doubles/floats to CSV, they'll now always have trailing zero.